Removed Extra modules
