--- sklearn/utils/modifiedunittest/__main__.py
@@ -1,12 +1,0 @@
-"""Main entry point"""
-
-import sys
-if sys.argv[0].endswith("__main__.py"):
-    sys.argv[0] = "python -m unittest"
-
-__unittest = True
-
-from .main import main, TestProgram, USAGE_AS_MAIN
-TestProgram.USAGE = USAGE_AS_MAIN
-
-main(module=None)
