--- sklearn/utils/modifiedunittest/test/dummy.py
@@ -1,1 +1,0 @@
-# Empty module for testing the loading of modules
